Field: Refactor constructor to not take cell array
This is the first of several commits that look for a refactor of the
code to decouple several things. As a first step we modify the
constructor of Field, such that instead of an array of cells it gets
a size and a list of coordinates where cells should be alive. For now,
the size parameter is not doing anything, but this should change on
a future commit.

Since we only needed an array of cells to initialize the Field, we can
now also stop exposing the Cells to binaries outside of the library.